Strengthen key derivation
This patch makes the derivation of the remote and local keys completely
independent even with a black box approach of pbkdf2.

In the previous configuration, if pbkdf2 was found to be reversible, then
an attacker with knowledge of the remote key could guess the local key.

The downside is that the master key's entropy is capped to 32 bytes
but that's more than any master password will ever be anyway so it's
not a real threat.